+++ CHANGELOG.md
@@ -42,6 +42,7 @@ The project versioning policy is now explicitly governed by SEMVER. See https://
 
 - Support async cancellations, ensuring that the connection pool is left in a clean state when cancellations occur. (#726)
 - The networking backend interface has [been added to the public API](https://www.encode.io/httpcore/network-backends). Some classes which were previously private implementation detail are now part of the top-level public API. (#699)
+- Add some connection pool events for trace extension. (#702)
 - Graceful handling of HTTP/2 GoAway frames, with requests being transparently retried on a new connection. (#730)
 - Add exceptions when a synchronous `trace callback` is passed to an asynchronous request or an asynchronous `trace callback` is passed to a synchronous request. (#717)
 - Drop Python 3.7 support. (#727)